new calss scorecard created
